Feat: Issue #744: Curl_trait created, connector curl separated and flow_curl created
